The expiry date is in UTC format, but for humans, on the index page, have made it local timezone.
It was a bit striking on the graph that the day could go to the next day with the local timezone / UCT difference so I've played down that by removing the day.

The &sup didn't always work ont he main page so put in <sup>*</sup> but in fact I had missed out the copyright line it referred to.